Move OneDimGridFactory under Numerics
